backend/requirements.txt
# Django
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3

# Database
psycopg2-binary==2.9.9

# Celery
celery==5.4.0
redis==5.0.8

# Utils
python-dotenv==1.0.1
ipaddress==1.0.23

# Dev
pytest==8.3.3
pytest-django==4.9.0
black==24.8.0
isort==5.13.2
